optimizado base de datos

+++ index.js
@@ -29,7 +29,7 @@ function format_num(n) {
 }
 
 function download_file(url, localFile, callback) {
-	//console.log('Start download : ' + url);
+	console.log('Start download : ' + url);
 	var localStream = fs.createWriteStream(localFile);
 	var out = request({
 		uri: url
@@ -61,13 +61,13 @@ function proces_file_save() {
 	var day, hour = '';
 	var check_hour = true;
 	handler.on('way', function(way) {
-		if (check_hour) {
-			//console.log(way.timestamp);
-			var date = new Date(parseFloat(way.timestamp) * 1000);
-			hour = date.getHours();
-			day = date.getUTCFullYear() + '-' + complete_date(parseInt(date.getUTCMonth()) + 1) + '-' + complete_date(date.getUTCDate()) + '-' + complete_date(date.getHours());
-			check_hour = false;
-		}
+		//if (check_hour) {
+		//console.log(way.timestamp);
+		var date = new Date(parseFloat(way.timestamp) * 1000);
+		//hour = date.getHours();
+		day = date.getUTCFullYear() + '-' + complete_date(parseInt(date.getUTCMonth()) + 1) + '-' + complete_date(date.getUTCDate()) + '-' + complete_date(date.getHours());
+		//check_hour = false;
+		//}
 		if (typeof way.tags().highway !== 'undefined' && osm_users.indexOf(way.user) !== -1) { //evalua las calles	
 			if (way.version === 1) {
 				++count[way.user].way.highways.v1;
@@ -78,17 +78,43 @@ function proces_file_save() {
 	});
 
 	reader.apply(handler);
+
 	db.serialize(function() {
-		var stmt = db.prepare("INSERT INTO osm_highway VALUES (?,?,?,?,?)");
-		for (var i = 0; i < osm_users.length; i++) {
-			stmt.run(name_directory + '-' + name_file, i + 1, day, count[osm_users[i]].way.highways.v1, count[osm_users[i]].way.highways.vx);
-		};
-		stmt.finalize();
+		var id_date = 0;
+		var stmt_date = db.prepare("INSERT INTO osm_date(osm_file,osm_date) VALUES (?,?)");
+		stmt_date.run(name_directory + '-' + name_file, day);
+		stmt_date.finalize();
+		db.each("SELECT MAX(id_date) as id_date  FROM osm_date", function(err, row) {
+			id_date = row.id_date;
+		}, function() {
+			var stmt_highway = db.prepare("INSERT INTO osm_highway VALUES (?,?,?,?)");
+			for (var i = 0; i < osm_users.length; i++) {
+				stmt_highway.run(i + 1, id_date, count[osm_users[i]].way.highways.v1, count[osm_users[i]].way.highways.vx);
+			};
+			stmt_highway.finalize();
+		});
 	});
-	var tempFile = fs.openSync(osm_file, 'r');
-	fs.closeSync(tempFile);
-	fs.unlinkSync(osm_file);
-	console.log('Remove file :' + osm_file);
+
+
+	//var tempFile = fs.openSync(osm_file, 'r');
+	//fs.closeSync(tempFile);
+	//fs.unlinkSync(osm_file);
+	//console.log('Remove file :' + osm_file);
+
+	//repite
+	if (!fs.exists(osm_file)) {
+		var tempFile = fs.openSync(osm_file, 'r');
+		fs.closeSync(tempFile);
+		fs.unlinkSync(osm_file);
+		console.log('Remove file :' + osm_file);
+	} else {
+		console.log('Error in remove file');
+	}
+
+	//url_file = get_url_file();
+	//osm_file = name_file + '.osc'
+	//download_file(url_file, osm_file, proces_file_save);
+
 }
 
 function get_url_file() {
@@ -122,7 +148,7 @@ function complete_date(str) {
 var url = 'http://planet.openstreetmap.org/replication/hour/000';
 var name_file = '';
 var num_file = 102;
-var num_directory = 18;
+var num_directory = 17;
 var name_directory = ''
 name_directory = '0' + num_directory;
 
@@ -130,4 +156,5 @@ setInterval(function() {
 	var url_file = get_url_file();
 	osm_file = name_file + '.osc'
 	download_file(url_file, osm_file, proces_file_save);
-}, 60* 60 * 1000);
+
+}, 2 * 60 * 1000);
